==> Code/zarni/04_GISMap.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

Libraries
```{r, warning = FALSE}
library(readxl)
library(data.table)
library(lubridate) #Date Time Object
library(ggplot2)
library(stringr) #Regular expression string splitting
library(rio)
library(MASS)
library(tigris)
library(sp)
library(broom)
library(dplyr)
library(rgdal)
```

UNUSED: BLOCKGROUP METHOD
Getting Census Block Group Shape files from Tigiris
```{r}
#MA <- block_groups("MA")
```

To make it plottable on ggplot
```{r}
# MA@data$id <- rownames(MA@data)
# MA.points <- fortify(MA, region = "id")
# MA.df <- MA.points %>% left_join(MA@data, by = c("id", "id"))
```

plotting base map in ggplot
```{r}
# ggplot() + 
#   geom_polygon(data = MA.df, aes(x = long, y = lat, group = group, fill = NA), color = 'black', fill = NA, size = 0.5)
```
END UNUSED: BLOCKGROUP METHOD


Methods using:

https://andrewbtran.github.io/NICAR/2017/maps/mapping-census-data.html
http://www.kevjohnson.org/making-maps-in-r/

ShapeFile of MA by Lower House boundaries
```{r}
MA_LH <- readOGR("../../Data/Shapefiles/", layer = "cb_2016_25_sldl_500k")
MA_LH@data$id <- rownames(MA_LH@data)
MA_LH.points <- fortify(MA_LH, region = "id")
MA_LH_data <- MA_LH.points %>% left_join(MA_LH@data, by = c("id", "id"))
```

Read in the Long and Lat file for T21 for MA
```{r}
MA_Lon_Lat <- import("../../Data/MA21_lonlat.csv")
View(MA_Lon_Lat)
coords <- MA_Lon_Lat[, c("lon", "lat")]
#Check to see if we have any missing
coords <- coords[complete.cases(coords),]
#Convert it to Spatial Points
sp <- SpatialPoints(coords)
```




```{r}
p <- ggplot() +
    geom_polygon(data = MA_LH_data, aes(x = long, y = lat, group = group), color = "black", size = 0.25, alpha = 0.25, fill = "ivory4")
p <- p +geom_point(data = MA_Lon_Lat, aes(x = lon, y = lat, color = "red"), size = 0.5 )
p <- p + coord_map()
p <- p + labs(x=NULL, y=NULL, 
                  title="T21 cities in MA",
                  subtitle=NULL,
                  caption="Source: T21 website")
p <- p + theme(plot.title=element_text(face="bold", family="Arial", size=20))
p <- p + theme(plot.caption=element_text(face="bold", family="Arial", size=10, color="gray", margin=margin(t=10, r=80)))
p <- p + theme(legend.position="none")
p <- p + theme(axis.line =  element_blank(),
                   axis.text =  element_blank(),
                   axis.ticks =  element_blank(),
                   panel.grid.major = element_blank(),
                   panel.grid.minor = element_blank(),
                   panel.border = element_blank(),
                   panel.background = element_blank()) 
print(p)
ggsave(p, file = "../../Data/MAT21.png", type = "cairo-png")

```
