==> Code/zarni/05_ModelData.Rmd <==
---
title: "R Notebook"
output:
  html_notebook: default
  pdf_document: default
---
Libraries

```{r, warning = FALSE}
library(readxl)
library(data.table)
library(lubridate) #Date Time Object
library(ggplot2)
library(stringr) #Regular expression string splitting
library(rio)
library(MASS)
library(tigris)
library(sp)
library(broom)
library(dplyr)
library(rgdal)
library(stargazer)
library(stats) #Principal Component Package
library(lme4)
library(lmerTest)
```

T21 External Raw Data 
```{r}
T21External <- import("../../Data/MACTownHealth21_v6.csv")
View(T21External)
```

Remove the 2 cases and remove the years before 2009 and removing 2018 (Data needs to be reformatted again)
```{r}
T21ExtSub <- T21External[T21External$Event != 2, ]
T21ExtSub <- T21ExtSub[-1]
T21ExtSub <- subset(T21ExtSub, Year > 2010)
T21ExtSub <- subset(T21ExtSub, Year < 2018)
View(T21ExtSub)

write.csv(T21ExtSub, "../../Data/T21ExtSub.csv")
View(T21ExtSub)
```

All T21 Imports

```{r}
AllT21TopRatios <- import("../../Data/AllT21TopRatios.csv")
View(AllT21TopRatios)
AllT21TopRatios <- AllT21TopRatios[-1]
View(AllT21TopRatios)
```

Merging the data file
```{r}
T21Merged <- T21ExtSub %>% left_join(AllT21TopRatios, by = c("Name" ="CityTown", "Year" = "year"))
```

```{r}
names(T21Merged) <- gsub(" ","",names(T21Merged))
```

```{r}
str(T21Merged)
T21Merged$Population <- as.numeric (T21Merged$Population)
write.csv(T21Merged, "../../Data/T21Mergedsecondattempt.csv")
```

```{r}
demo <- import("../../Data/T21Mergedsecondattempt.csv")
View(demo)
dim(demo)
```



```{r}
T21Merged$Event <- as.numeric(T21Merged$Event)
```



Logit Model
```{r}
logitmodel1 <- glm(Event ~  Population + NearbyT21Ratio + PhysicallyUnhealthyDays + MentallyUnhealthyDays + PercentSmokers + PercentExcessiveDrinking + TeenBirthRate + ChlamydiaRate + PercentUnemployed + ViolentCrimeRate   , family = binomial(link = "logit"), data = T21Merged )

summary(logitmodel1)
```

Table format
```{r}
stargazer(logitmodel, type = "text")

```

```{r}
exp(coefficients(logitmodel))
```
Should NearbyT21 Ratio be coded as a factor?

```{r}
T21Merged$FNT21Ratio <- as.factor(T21Merged$NearbyT21Ratio)
```

Logit Model 2
```{r}
logitmodel2 <- glm(Event ~  Population + FNT21Ratio + PhysicallyUnhealthyDays + MentallyUnhealthyDays + PercentSmokers + PercentExcessiveDrinking + TeenBirthRate + ChlamydiaRate + PercentUnemployed + ViolentCrimeRate   , family = binomial(link = "logit"), data = T21Merged )

summary(logitmodel2)
```


```{r}
stargazer(logitmodel2, type = "text")
```
```{r}
exp(coefficients(logitmodel2))
```

**Principal Component Analysis**


Principal Component Analysis

```{r}
pca_test <- T21Merged
```

```{r}
View(pca_test)
names(pca_test)
pca_test_01 <- pca_test[,c(7:36)]
```

What is the difference between using covariance matrix and correlation matrix?

```{r}
pca_job <- princomp(pca_test_01, cor = FALSE)
```

### Master File Latest Model Building

```{r}
master <- rio::import("../../Data/master.csv")
#View(master)
#names(master)
#Dropping Chlamydia Rate becasue we do NOT understand the measurement for now.
master <- subset(master, select = -c(25))
master <- master[,-c(1:2)]
View(master)
```

### Getting the Top 10 ratios
```{r}
AllT21TopRatios <- import("../../Data/AllT21TopRatios.csv")
#View(AllT21TopRatios)
AllT21TopRatios <- AllT21TopRatios[-1]
#View(AllT21TopRatios)
```

### Merging Ratios and the Master File for 2 Year Lags
```{r}
T21_TopRatios_Master_lag_2 <- master %>% left_join(AllT21TopRatios, by = c("Name.x" ="CityTown", "year_dummy-2" = "year"))
names(T21_TopRatios_Master_lag_2)[73] <- c("nearbyt21ratio")
```

```{r}
T21_TopRatios_Master_lag_2 <- T21_TopRatios_Master_lag_2[T21_TopRatios_Master_lag_2$Event !=2,]
T21_TopRatios_Master_lag_2$Population <- as.numeric(T21_TopRatios_Master_lag_2$Population)
sort(names(T21_TopRatios_Master_lag_2))
```



```{r}
logitmodel_lag_21 <- glm(Event ~  Population + prct_r_2012 + prct_d_2012 + prct_d_2016 + prct_r_2016 + mayor_ind + nearbyt21ratio   , family = binomial(link = "logit"), data = T21_TopRatios_Master_lag_2)
summary(logitmodel_lag_21)
```

### Merging Ratios and the Master File for 1 Year Lags
```{r}
T21_TopRatios_Master_lag_1 <- master %>% left_join(AllT21TopRatios, by = c("Name.x" ="CityTown", "year_dummy-1" = "year"))
names(T21_TopRatios_Master_lag_1)[73] <- c("nearbyt21ratio")
```

```{r}
T21_TopRatios_Master_lag_1<- T21_TopRatios_Master_lag_2[T21_TopRatios_Master_lag_1$Event !=2,]
T21_TopRatios_Master_lag_1$Population <- as.numeric(T21_TopRatios_Master_lag_1$Population)
sort(names(T21_TopRatios_Master_lag_1))
```

```{r}
logitmodel_lag_11 <- glm(Event ~  Population + prct_r_2012 + prct_d_2012 + prct_d_2016 + prct_r_2016 + mayor_ind + nearbyt21ratio   , family = binomial(link = "logit"), data = T21_TopRatios_Master_lag_1)
summary(logitmodel_lag_11)
```

Neighborhood Pure Distance Method
```{r}
AllT21ratios <- rio:: import("../../Data/AllT21Ratios.csv")
AllT21ratios <- AllT21ratios[-1]
nrow(AllT21ratios)
length(unique(AllT21ratios$CityTown))
```

### 2 year Lag Joining

```{r}
#View(master)
T21_Ratios_Master_lag_2 <- master %>% left_join(AllT21ratios, by = c("Name.x" ="CityTown", "year_dummy-2" = "year"))
names(T21_Ratios_Master_lag_2)[73] <- c("nearbyt21ratio")
```

```{r}
T21_Ratios_Master_lag_2 <- T21_Ratios_Master_lag_2[T21_Ratios_Master_lag_2$Event !=2,]
T21_Ratios_Master_lag_2$Population <- as.numeric(T21_Ratios_Master_lag_2$Population)
sort(names(T21_Ratios_Master_lag_2))
```

Neighborhood Distance function Method 2
```{r}
logitmodel_lag_22 <- glm(Event ~  Population + prct_r_2012 + prct_d_2012 + prct_d_2016 + prct_r_2016 + mayor_ind + nearbyt21ratio   , family = binomial(link = "logit"), data = T21_Ratios_Master_lag_2)
summary(logitmodel_lag_22)
```

### 1 year lag joining
```{r}
#View(master)
T21_Ratios_Master_lag_1 <- master %>% left_join(AllT21ratios, by = c("Name.x" ="CityTown", "year_dummy-1" = "year"))
names(T21_Ratios_Master_lag_1)[73] <- c("nearbyt21ratio")
```

```{r}
T21_Ratios_Master_lag_1<- T21_Ratios_Master_lag_1[T21_Ratios_Master_lag_1$Event !=2,]
T21_Ratios_Master_lag_1$Population <- as.numeric(T21_Ratios_Master_lag_1$Population)
sort(names(T21_Ratios_Master_lag_1))
```

Neighborhood Distance function Method 2
```{r}
logitmodel_lag_21 <- glm(Event ~  Population + prct_r_2012 + prct_d_2012 + prct_d_2016 + prct_r_2016 + mayor_ind + nearbyt21ratio   , family = binomial(link = "logit"), data = T21_Ratios_Master_lag_1)
summary(logitmodel_lag_21)

```

###Getting the Contigency based ratio which is already 1 year lagged

```{r}
AllT21ContingencyRatios <- rio::import("../../Data/borderratios_updated.csv")
#View(AllT21ContingencyRatios)
#nrow(AllT21ContingencyRatios)
AllT21ContingencyRatios <- AllT21ContingencyRatios[-1]
AllT21ContigencyRatios$ratio <- as.numeric(AllT21ContigencyRatios$ratio)
#AllT21ContigencyRatios$ratio
```

```{r}
nrow(master)
nrow(AllT21ContingencyRatios)
#names(AllT21ContingencyRatios)

master$Name.x <- tolower(master$Name.x)
#master$Name.x
#match(unique(master$Name.x), unique(AllT21ContingencyRatios$TownName))
#match(unique(master$Year), unique(AllT21ContingencyRatios$year))

T21_Contingency_Master <- master %>% left_join(AllT21ContingencyRatios, by = c("Name.x" = "TownName", "Year" = "year")) 
#View(T21_Contingency_Master)

```

```{r}
T21_Contingency_Master <- T21_Contingency_Master[T21_Contingency_Master$Event !=2,]
T21_Contingency_Master$Population <- as.numeric(T21_Contingency_Master$Population)
#sort(names(T21_Contingency_Master))
#T21_Contingency_Master$ratio
T21_Contingency_Master$ratio <- as.numeric(T21_Contingency_Master$ratio)
```


Neighborhood Contingency
```{r}
logitmodel_lag_13 <- glm(Event ~  Population + prct_r_2012 + prct_d_2012 + prct_d_2016 + prct_r_2016 + mayor_ind + ratio   , family = binomial(link = "logit"), data = T21_Contingency_Master)
summary(logitmodel_lag_13)
#View(T21_Contigency_Master)
```


### Contigency based 2 years
```{r}
AllT21ContingencyRatios <- rio::import("../../Data/borderratios_updated.csv")
#View(AllT21ContingencyRatios)
#nrow(AllT21ContingencyRatios)
AllT21ContingencyRatios <- AllT21ContingencyRatios[-1]
AllT21ContigencyRatios$ratio <- as.numeric(AllT21ContigencyRatios$ratio)
#AllT21ContigencyRatios$ratio
```

```{r}
nrow(master)
nrow(AllT21ContingencyRatios)
#names(AllT21ContingencyRatios)

master$Name.x <- tolower(master$Name.x)
#master$Name.x
#match(unique(master$Name.x), unique(AllT21ContingencyRatios$TownName))
#match(unique(master$Year), unique(AllT21ContingencyRatios$year))

T21_Contingency_Master <- master %>% left_join(AllT21ContingencyRatios, by = c("Name.x" = "TownName", "year_dummy-1" = "year")) 
#View(T21_Contingency_Master)

```

```{r}
T21_Contingency_Master <- T21_Contingency_Master[T21_Contingency_Master$Event !=2,]
T21_Contingency_Master$Population <- as.numeric(T21_Contingency_Master$Population)
#sort(names(T21_Contingency_Master))
#T21_Contingency_Master$ratio
T21_Contingency_Master$ratio <- as.numeric(T21_Contingency_Master$ratio)
```


Neighborhood Contingency
```{r}
logitmodel_lag_23 <- glm(Event ~  Population + prct_r_2012 + prct_d_2012 + prct_d_2016 + prct_r_2016 + mayor_ind + ratio   , family = binomial(link = "logit"), data = T21_Contingency_Master)
summary(logitmodel_lag_23)
#View(T21_Contigency_Master)
```
